pyproject.toml
[project]
name = "qa-guru-graduation-project-ui"
version = "0.0.1"
description = "graduation project for course qa guru python automation"
authors = [
    {name = "Falin Pavel", email = "falinpavel96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
poetry-core = "^2.1.3"
allure-pytest = "2.15.0"
allure-python-commons = "2.15.0"
pytest = "8.4.2"
python-dotenv = "1.1.1"
selene = "2.0.0rc9"
selenium = "4.35.0"
pytest-xdist = "3.8.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
